implements feature

+++ src/frontend/src/lib/utils/format.utils.ts
@@ -123,7 +123,7 @@ export const formatNanosecondsToTimestamp = (nanoseconds: bigint): number => {
 };
 
 export const formatToShortDateString = ({ date, i18n }: { date: Date; i18n: I18n }): string =>
-	date.toLocaleDateString(i18n?.lang ?? Languages.ENGLISH, { month: 'long' });
+	date.toLocaleDateString(i18n?.lang ?? Languages.ENGLISH, { month: 'short' });
 
 const getRelativeTimeFormatter = (language?: Languages) =>
 	new Intl.RelativeTimeFormat(language ?? Languages.ENGLISH, { numeric: 'auto' });
